requirements.txt
tweepy
python-dotenv
babel
num2words